refactor(transit-display): rename + extract Entry
Rename `TransitDisplayEntry2` -> `TransitDisplayEntry` (component +
Props) and move it out of `transit-displays-2.tsx` into its own
`transit-display-entry.tsx`. The previously hardcoded entry-only
size tables (`ROW_TEXT_CLASS_BY_SIZE`,
`TIMETABLE_ENTRY_ATTRIBUTES_LABELS_SIZE_BY_SIZE`,
`DISTANCE_BADGE_SIZE_BY_SIZE`) move with the component; the caller
file (`transit-displays-2.tsx`) now imports the entry instead of
defining it and sheds 14 unused imports left over from the entry
body. Stories file is renamed (`transit-display-entry-2.stories.tsx`
-> `transit-display-entry.stories.tsx`) and its imports / Storybook
title / component refs are updated to match.
